requirements.txt
streamlit
Pillow
pytesseract
pdf2image
google-generativeai
easyocr
numpy 